FSM graphs and process pseudocode
This closes #29 and #28.

The graph representation for the e finite state machines
as described in `players.py`.

Also for representing the process two pieces of pseudocode.
One for the data collection and one for the moran process.

The pseudocodes are in a tex file, not sure where you would like
them.

Update gitignore to ignore pdf of tex files.

Ashlock's fortress strategies for comparison are also included.n

Eating a letter