Write a first draft.
This has removed a number of things to make the paper a bit more
efficient. I have taken Marc's text about training strategies and moved
it around to fit better. It is also one of the main narratives of the
story now.

Note that I've removed "we" and "our" throughout.